Improved and updated dotbot install configuration file
Rearranged shell executions in order for the setup file to execute prior to the config file cleaning or the linking occuring. Fixed path export command. Added link to the dotfiles folder. Improved the clean command to clean only config folders and not the entire home directory. Fixed the discocss linking.