Added latest version of scope.sh for ranger
Added the latest version of ranger's scope.sh file which defines all of the file previews for ranger and implemented it in the ranger config file. Also modified the dotbot install configuration file to create a link for the whole folder and not just the ranger config file since the ranger directory now contains more than just that single ranger config.